projet: ProjectCard component finished
